Block to handle SIGINT signal and longer delay
When the script is running as a service on RPi, it is killed via `systemctl stop` which sends a SIGINT signal. Since it was not handled properly, gpio pins were not reset. Also, added a 5 second delay in between sensor reading as 0.5 second interval was just absurd. We are not trading stocks here.